Move BuildArtifact to common
The BuildArtifact is used in both the master and the slave code so
this just changes the file location to reflect that.
